Move client-side rendering code out of MathHelper
